code improvement duplicates AbstractMdSchemaResponseRowXml Signed-off-by: dbulahov <bulahovdenis@gmail.com>
