+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla_rowset/mdschema/MdSchemaActionsResponseRowXml.java
@@ -18,7 +18,6 @@ import java.io.Serializable;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.ActionTypeEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.CoordinateTypeEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.InvocationEnum;
-import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla_rowset.Row;
 
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
@@ -31,28 +30,10 @@ import jakarta.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "MdSchemaActionsResponseRowXml")
-public class MdSchemaActionsResponseRowXml extends Row implements Serializable {
+public class MdSchemaActionsResponseRowXml extends AbstractMdSchemaResponseRowXml implements Serializable {
 
     @XmlTransient
     private static final long serialVersionUID = 8909756522012364431L;
-
-    /**
-     *  The name of the database.
-     */
-    @XmlElement(name = "CATALOG_NAME", required = false)
-    private String catalogName;
-
-    /**
-     *  The name of the schema.
-     */
-    @XmlElement(name = "SCHEMA_NAME", required = false)
-    private String schemaName;
-
-    /**
-     *  The name of the cube.
-     */
-    @XmlElement(name = "CUBE_NAME", required = false)
-    private String cubeName;
 
     /**
      *  The name of this action.
@@ -136,30 +117,6 @@ public class MdSchemaActionsResponseRowXml extends Row implements Serializable {
      */
     @XmlElement(name = "INVOCATION", required = false)
     private InvocationEnum invocation;
-
-    public String getCatalogName() {
-        return catalogName;
-    }
-
-    public void setCatalogName(String catalogName) {
-        this.catalogName = catalogName;
-    }
-
-    public String getSchemaName() {
-        return schemaName;
-    }
-
-    public void setSchemaName(String schemaName) {
-        this.schemaName = schemaName;
-    }
-
-    public String getCubeName() {
-        return cubeName;
-    }
-
-    public void setCubeName(String cubeName) {
-        this.cubeName = cubeName;
-    }
 
     public String getActionName() {
         return actionName;
