+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla_rowset/mdschema/MdSchemaDimensionsResponseRowXml.java
@@ -17,7 +17,6 @@ import java.io.Serializable;
 
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.DimensionTypeEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.DimensionUniqueSettingEnum;
-import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla_rowset.Row;
 
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
@@ -30,28 +29,10 @@ import jakarta.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "MdSchemaDimensionsResponseRowXml")
-public class MdSchemaDimensionsResponseRowXml extends Row implements Serializable {
+public class MdSchemaDimensionsResponseRowXml extends AbstractMdSchemaResponseRowXml implements Serializable {
 
     @XmlTransient
     private static final long serialVersionUID = 474404709256189241L;
-
-    /**
-     * The name of the database.
-     */
-    @XmlElement(name = "CATALOG_NAME", required = false)
-    private String catalogName;
-
-    /**
-     * The name of the schema.
-     */
-    @XmlElement(name = "SCHEMA_NAME", required = false)
-    private String schemaName;
-
-    /**
-     * The name of the cube.
-     */
-    @XmlElement(name = "CUBE_NAME", required = false)
-    private String cubeName;
 
     /**
      * The name of the dimension.
@@ -163,30 +144,6 @@ public class MdSchemaDimensionsResponseRowXml extends Row implements Serializabl
      */
     @XmlElement(name = "DIMENSION_IS_VISIBLE", required = false)
     private Boolean dimensionIsVisible;
-
-    public String getCatalogName() {
-        return catalogName;
-    }
-
-    public void setCatalogName(String catalogName) {
-        this.catalogName = catalogName;
-    }
-
-    public String getSchemaName() {
-        return schemaName;
-    }
-
-    public void setSchemaName(String schemaName) {
-        this.schemaName = schemaName;
-    }
-
-    public String getCubeName() {
-        return cubeName;
-    }
-
-    public void setCubeName(String cubeName) {
-        this.cubeName = cubeName;
-    }
 
     public String getDimensionName() {
         return dimensionName;
