+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla_rowset/mdschema/MdSchemaLevelsResponseRowXml.java
@@ -20,7 +20,6 @@ import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.LevelDbTypeEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.LevelOriginEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.LevelTypeEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.LevelUniqueSettingsEnum;
-import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla_rowset.Row;
 
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
@@ -33,29 +32,10 @@ import jakarta.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "MdSchemaLevelsResponseRowXml")
-public class MdSchemaLevelsResponseRowXml extends Row implements Serializable {
+public class MdSchemaLevelsResponseRowXml extends AbstractMdSchemaResponseRowXml implements Serializable {
 
     @XmlTransient
     private static final long serialVersionUID = 1478136849363120910L;
-
-    /**
-     * The name of the database.
-     */
-    @XmlElement(name = "CATALOG_NAME", required = false)
-    private String catalogName;
-
-
-    /**
-     * The name of the schema.
-     */
-    @XmlElement(name = "SCHEMA_NAME", required = false)
-    private String schemaName;
-
-    /**
-     * The name of the cube.
-     */
-    @XmlElement(name = "CUBE_NAME", required = false)
-    private String cubeName;
 
     /**
      * The unique name of the
@@ -412,30 +392,6 @@ public class MdSchemaLevelsResponseRowXml extends Row implements Serializable {
      */
     @XmlElement(name = "LEVEL_ORIGIN", required = false)
     private LevelOriginEnum levelOrigin;
-
-    public String getCatalogName() {
-        return catalogName;
-    }
-
-    public void setCatalogName(String catalogName) {
-        this.catalogName = catalogName;
-    }
-
-    public String getSchemaName() {
-        return schemaName;
-    }
-
-    public void setSchemaName(String schemaName) {
-        this.schemaName = schemaName;
-    }
-
-    public String getCubeName() {
-        return cubeName;
-    }
-
-    public void setCubeName(String cubeName) {
-        this.cubeName = cubeName;
-    }
 
     public String getDimensionUniqueName() {
         return dimensionUniqueName;
