+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla_rowset/mdschema/MdSchemaPropertiesResponseRowXml.java
@@ -20,7 +20,6 @@ import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.PropertyCardinalityEn
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.PropertyContentTypeEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.PropertyOriginEnum;
 import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.enums.PropertyTypeEnum;
-import org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla_rowset.Row;
 
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
@@ -33,29 +32,10 @@ import jakarta.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "MdSchemaPropertiesResponseRowXml")
-public class MdSchemaPropertiesResponseRowXml extends Row implements Serializable {
+public class MdSchemaPropertiesResponseRowXml extends AbstractMdSchemaResponseRowXml implements Serializable {
 
     @XmlTransient
     private static final long serialVersionUID = 8022966927923698289L;
-
-    /**
-     * The name of the database.
-     */
-    @XmlElement(name = "CATALOG_NAME", required = false)
-    private String catalogName;
-
-
-    /**
-     * The name of the schema.
-     */
-    @XmlElement(name = "SCHEMA_NAME", required = false)
-    private String schemaName;
-
-    /**
-     * The name of the cube.
-     */
-    @XmlElement(name = "CUBE_NAME", required = false)
-    private String cubeName;
 
     /**
      * The unique name of the dimension.
@@ -326,30 +306,6 @@ public class MdSchemaPropertiesResponseRowXml extends Row implements Serializabl
      */
     @XmlElement(name = "PROPERTY_IS_VISIBLE", required = false)
     private Boolean propertyIsVisible;
-
-    public String getCatalogName() {
-        return catalogName;
-    }
-
-    public void setCatalogName(String catalogName) {
-        this.catalogName = catalogName;
-    }
-
-    public String getSchemaName() {
-        return schemaName;
-    }
-
-    public void setSchemaName(String schemaName) {
-        this.schemaName = schemaName;
-    }
-
-    public String getCubeName() {
-        return cubeName;
-    }
-
-    public void setCubeName(String cubeName) {
-        this.cubeName = cubeName;
-    }
 
     public String getDimensionUniqueName() {
         return dimensionUniqueName;
